Fixing #i rendering
On the docs page, the letter `i` is not rendered for nuget source snippets. This might be because the language is set to `fsharp`. 

The NuGet installation snippets don't have `fsharp` as the language of the code snippet and the letter `r` renders correctly. This is a quick fix and might need more investigation on the F# Formatting side.