Binder button fix
The "Run in Binder" button from the main page is currently broken. I believe the problem can be traced to the Dockerfile used to build the Binder environment in the /docs folder, but I can't test this locally. I recommend installing the latest stable tool from Nuget (instead of the latest build) which should fix the problem. Better to have a working Binder environment than a broken one.